Basic functionality to handle reaction requests in desktop code
The desktop code handling serial communications now has the basic
functionality to handle reaction requests, using the IDs (2-char)
of the microbits involved, and a table of accepted reactions.
This has been tested by giving fake requests, but not on actual
requests coming from the hub. The functionality to send these
requests has to still be added to the MBs and the hub.

Committed the reaction table "elementsDB-v3.csv", and a sample
scenario file (scenario.csv).

Also made changes to the hub and MB codes in order for everything
to work together with the updated serial_communication.py