Reactions work otherwise, but updated symbols don't reach the MBs
The data on the new symbols that should update the MBs gets
corrupted between the USB and the hub, and thus gets rejected
at the MBs as incorrect data. The reactions are updated to the
table on the desktop system, but not onto the microbits.

A possible solution is to do a check on the message integrity
on the hub, requesting the data again if its not correct before
sending it onward to the MBs.