Committing a Windows-edited version of scenario_00.csv that works correctly
The previous version of scenario_00.csv was causing the program to hang
for some reason. It is likely that it had something to do with different
encoding in the lines that were edited using a linux machine, as the
scenario files had been made in Windows. This error was consistent.
The new version, built from a backup and only edited in Windows, appears
to work correctly.