Added new scenarios, and fixed a bug in Scenario.py
Added Kenny's new scenario files, and updated scenario.csv

Scenario.py had a bug that was causing the first microbit in the
master_table to not get its element correctly, and made the program
hang. This was due to an if condition that was incorrectly failing
with the first row, as its index is 0, and the resulting list, [0],
evaluated to false. This is now fixed, checking for the size of the
list instead.